requirements.txt
streamlit
diffusers
transformers
accelerate
safetensors
torch
torchvision
pillow
numpy
tqdm
